Added cuisine type to url param
Added the cuisine type to url params and ensured when page can be navigated with browser back/forward buttons. Also ensure that the search bar reflects the state of the searchParams.. e.g. If ?search=hello the searchbar value should be hello. If ?search= then the search bar should be empty. This is the same concept with the selected cuisine type.