Restyled/Updated Searchbar filter dropdowns
Restyles the searchbar filters and added a close button on them so the user can easily remove a filter. Restyling in progress to ensure when the user types in the name filter it resizes appropriately. Also updated the filter dropdowns for category/rating as they werent going to work how they were before due to overflow-x: scroll; overflow-y: visible issues